requirements.txt
streamlit>=1.32.0
groq>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
PyPDF2>=3.0.1
python-docx>=1.1.0
bcrypt>=4.1.2
streamlit-mic-recorder>=0.0.8
